Support spark in query_data_parametric
Summary: As titled. Otherwise for very large datasets we see the Presto memory limit error.

Reviewed By: j-jiafei

Differential Revision: D29020301

fbshipit-source-id: a35198cf0da83f2fc454e92844d6a7ea17e2b8f7

+++ reagent/preprocessing/transforms.py
@@ -116,7 +116,7 @@ class DenseNormalization:
             value, presence = value.to(self.device), presence.to(self.device)
             presence[torch.isnan(value)] = 0
             value[torch.isnan(value)] = 0
-            data[k] = self._preprocessor(value, presence)
+            data[k] = self._preprocessor(value, presence).float()
 
         return data
